# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for the e-commerce application",
        "version": "1.0.0",
        "title": "E-Commerce API",
        "contact": {
            "email": "dilipkumarbc8055@gmail.com"
        }
    },
    "host": "localhost:8080",
    "securityDefinitions": {
        "JWT": {
            "in": "header",
            "name": "Authorization",
            "type": "apiKey"
        }
    },
    "basePath": "/",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/api/product": {
            "get": {
                "tags": [
                    "Product"
                ],
                "security": [
                    {
                        "JWT": {}
                    }
                ],
                "summary": "Get a list of products",
                "responses": {
                    "200": {
                        "description": "A list of products",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Product"
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Create a new product",
                "security": [
                    {
                        "JWT": {}
                    }
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "product",
                        "description": "Product object to create",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Product"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Product created successfully"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Product": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "name": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "price": {
                    "type": "number",
                    "format": "double"
                }
            },
            "required": [
                "id",
                "name",
                "price"
            ]
        }
    }
}